+++ src/query/remote.tsx
@@ -1,4 +1,4 @@
-import { QueryEngine } from "@comunica/query-sparql-rdfjs";
+import { QueryEngine } from "@comunica/query-sparql";
 import { IDataSource } from "@comunica/types";
 import { fromRDF, compact, flatten } from "jsonld";
 import { context, endpoint } from "../config";
@@ -14,7 +14,8 @@ export function format_query(query: string) {
     return `${prefix}\n${query}`;
 }
 
-export async function query_select(query: string, engine: QueryEngine, sources = defaultSource) {
+export async function query_select(query: string, sources = defaultSource) {
+    const engine = new QueryEngine()
     const response = await engine.queryBindings(
         format_query(query),
         { sources: sources, unionDefaultGraph: true }
@@ -23,7 +24,8 @@ export async function query_select(query: string, engine: QueryEngine, sources =
     return bindings;
 }
 
-export async function query_quads(query: string, engine: QueryEngine, sources = defaultSource) {
+export async function query_quads(query: string, sources = defaultSource) {
+    const engine = new QueryEngine()
     const response = await engine.queryQuads(
         format_query(query),
         { sources: sources, unionDefaultGraph: true }
@@ -32,16 +34,16 @@ export async function query_quads(query: string, engine: QueryEngine, sources = 
     return quads;
 }
 
-export async function query_jsonld(query: string, engine: QueryEngine, sources = defaultSource): Promise<Array<object>> {
-    const quads = query_quads(query, engine, sources)
+export async function query_jsonld(query: string, sources = defaultSource): Promise<Array<object>> {
+    const quads = await query_quads(query, sources)
     const document = await fromRDF(quads);
     const document_flat = await flatten(document, context);
     const document_compact = await compact(document_flat, context, {compactArrays: false});
     return Array.isArray(document_compact["@graph"]) ? document_compact["@graph"] : [];
 }
 
-export async function query_single(query: string, engine: QueryEngine, sources = defaultSource) {
-    const bindings = await query_select(query, engine, sources)
+export async function query_single(query: string, sources = defaultSource) {
+    const bindings = await query_select(query, sources)
     const binding =  bindings.find(Boolean)
     if (binding) {
         for (const node of binding.values()) {
